Refactor lead generation code and update API calls
This commit refactors the lead generation code in the `generate_leads` function. It removes the commented out code for generating leads from Reddit and updates the API calls for generating leads from Twitter and LinkedIn using Social Searcher. The result data is now saved to a JSON file with the keys `twitter_leads` and `linkedin_leads`. The response format for Twitter leads is also modified to include additional information such as the username, user ID, and timestamp of the tweet. This improves the efficiency and accuracy of lead generation.